moved Generate Colour Array into helper
