+++ apps/sudoku-backend/src/app/sudoku/helpers/sudoku-solver.ts
@@ -375,6 +375,68 @@ export function solveRegionSudoku(sudoku: number[], colours: number[], type: str
   return coloursudoku;
 }
 
+/***
+ * Creates assignment of colors based on a given sudoku array
+ * @param sudokuValueArray the array containing the actual values
+ * @private
+ */
+export function solveDiaColourSudoku(sudokuValueArray: number[]) {
+
+  // clone the array with the values for later remembering
+  const takenValues = [...sudokuValueArray];
+  // create an empty array for the colors to save and later return
+  const colorArray = Array(sudokuValueArray.length);
+
+  // for every color idx (1-9)
+  for (let colorIdx = 1; colorIdx < 10; colorIdx++) {
+    // get 9 values for each color
+    for (let valueToLookFor = 1; valueToLookFor < 10; valueToLookFor++) {
+
+      // get a random number to start the index
+      let currentIdxInValueArray = getRandomInt(0, 80);
+
+      // search as long as a fitting value is found
+      let isNoValueFound = true;
+      while (isNoValueFound) {
+
+        // if the idx reaches the end, start from beginning of array
+        if (currentIdxInValueArray >= 81) {
+          currentIdxInValueArray = 0;
+        }
+
+        if (takenValues[currentIdxInValueArray] == valueToLookFor) {
+          // the value is one that we look for
+          // stop iteration
+          isNoValueFound = false;
+          // assign color to the result
+          colorArray[currentIdxInValueArray] = colorIdx;
+          // mark the value as taken by assigning a -1
+          takenValues[currentIdxInValueArray] = -1;
+        }
+
+        // next iteration search on next value
+        currentIdxInValueArray += 1;
+
+      }
+    }
+  }
+  console.log(colorArray);
+  return colorArray;
+}
+
+/**
+ * Returns a random integer between min (inclusive) and max (inclusive).
+ * The value is no lower than min (or the next integer greater than min
+ * if min isn't an integer) and no greater than max (or the next integer
+ * lower than max if max isn't an integer).
+ * Using Math.round() will give you a non-uniform distribution!
+ */
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
 
 export function solveColourSudoku(sudoku: number[], colours: number[], type: string) {
   const saved = new Array<Array<Array<number>>>();
